Fixed type for -dr
It was set to float by mistake. It's fixed now.
Resolved #209

+++ honeybee/radiance/parameters/gridbased.py
@@ -142,7 +142,7 @@ class GridBasedParameters(AdvancedRadianceParameters):
         """
 
         self.add_radiance_number('dr', descriptive_name='direct relays',
-                                 num_type=float, attribute_name='direct_sec_relays')
+                                 num_type=int, attribute_name='direct_sec_relays')
         self.direct_sec_relays = None
         """
         -dr N
@@ -365,6 +365,7 @@ class GridBasedParameters(AdvancedRadianceParameters):
         """
         return {"gridbased_parameters": self.to_rad_string()
                 }
+
 
 class LowQuality(GridBasedParameters):
     """Radiance parmaters for a quick analysis."""
